+++ src/gateway/converter/sql_to_substrait.py
@@ -1,8 +1,8 @@
 # SPDX-License-Identifier: Apache-2.0
 """Routines to convert SparkConnect plans to Substrait plans."""
 from backends.backend import Backend
-from gateway.converter.add_extension_uris import AddExtensionUris
 from substrait.gen.proto import plan_pb2
+from transforms.add_extension_uris import AddExtensionUris
 
 
 def convert_sql(backend: Backend, sql: str) -> plan_pb2.Plan:
