+++ src/transforms/simplify_casts.py
@@ -2,10 +2,10 @@
 """A library to search Substrait plan for local files."""
 from typing import Any
 
-from gateway.converter.output_field_tracking_visitor import get_plan_id
 from gateway.converter.symbol_table import SymbolTable
 from substrait.gen.proto import algebra_pb2
 from substrait_visitors.substrait_plan_visitor import SubstraitPlanVisitor
+from transforms.output_field_tracking_visitor import get_plan_id
 
 
 # pylint: disable=no-member,fixme
